+++ CHANGELOG.md
@@ -48,6 +48,11 @@ CHANGELOG
   Host. Logging can be enabled or disabled during host start-up by calling
   ServiceHost.setRequestLoggingInfo OR by making a PATCH to ServiceHostManagementService.
 
+* MigrationTaskService supports migrating deleted documents. When query spec includes 
+  "INCLUDE_DELETED" option, migration task first posts the document, then deletes
+  it in destination nodes.
+
+
 ## 1.5.4
 
 * Updates to migration task service to assert zero documentOwner mismatches